requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.0
frozenlist==1.7.0
h11==0.16.0
holidays==0.81
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.11.0
kiwisolver==1.4.9
Markdown==3.9
matplotlib==3.10.6
multidict==6.6.4
numpy==2.3.3
openai==0.28.0
packaging==25.0
pandas==2.3.2
phonenumbers==9.0.15
pillow==11.3.0
propcache==0.3.2
prophet==1.1.7
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stanio==0.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
psycopg2-binary==2.9.10
